backend/requirements.txt
asgiref==3.8.1
Django==5.1
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
mysqlclient==2.2.4
psycopg2-binary==2.9.9
PyJWT==2.9.0
python-dotenv==1.0.1
pytz==2024.1
sqlparse==0.5.1
python-decouple==3.8
celery==5.3.6
django-celery-results==2.5.1
africastalking==1.2.7
django-filter==23.5
